Move calculation functions out of builder.rs
